requirements.txt
# Web Framework
Flask==2.0.1
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1

# AWS
boto3==1.26.137
botocore==1.29.137
s3transfer==0.6.1

# Audio Processing
# Note: Install PyAudio separately
google-cloud-speech==2.21.0
pyttsx3==2.90

# AI/ML
openai==0.27.0
numpy==1.24.3
pandas==2.0.1

# Utilities
python-dotenv==0.19.0
requests==2.28.2
urllib3==1.26.15
six==1.16.0
protobuf==4.22.3

# Development
python-dateutil==2.8.2
pytz==2023.3

openpyxl==3.1.2